create task folder and move Task.js into it
